requirements.txt

aiohappyeyeballs==2.6.1 
aiohttp==3.12.13 
aiosignal==1.3.2 
altair==5.5.0 
annotated-types==0.7.0 
anyio==4.9.0 
attrs==25.3.0 
beautifulsoup4==4.13.4 
blinker==1.9.0 
blis==1.3.0 
cachetools==6.1.0 
catalogue==2.0.10 
certifi==2025.6.15 
charset-normalizer==3.4.2 
click==8.2.1 
cloudpathlib==0.21.1 
colorama==0.4.6 
confection==0.1.5 
cymem==2.0.11 
distro==1.9.0 
fastapi==0.115.13 
feedparser==6.0.11 
filelock==3.18.0 
frozenlist==1.7.0 
gitdb==4.0.12 
GitPython==3.1.44 
h11==0.16.0 
httpcore==1.0.9 
httpx==0.28.1 
idna==3.10 
Jinja2==3.1.6 
jiter==0.10.0 
jsonschema==4.24.0 
jsonschema-specifications==2025.4.1 
langcodes==3.5.0 
language_data==1.3.0 
lxml==5.4.0 
marisa-trie==1.2.1 
markdown-it-py==3.0.0 
MarkupSafe==3.0.2 
mdurl==0.1.2 
 
multidict==6.5.1 
murmurhash==1.0.13 
narwhals==1.44.0 
newsapi-python==0.2.7 
numpy==2.3.1 
openai==1.91.0 
packaging==25.0 
pandas==2.3.0 
pillow==11.2.1 
plotly==6.1.2 
preshed==3.0.10 
propcache==0.3.2 
protobuf==6.31.1 
pyarrow==20.0.0 
pydantic==2.11.7 
pydantic_core==2.33.2 
pydeck==0.9.1 
Pygments==2.19.2 
PySocks==1.7.1 
python-dateutil==2.9.0.post0 
python-dotenv==1.1.1 
python-multipart==0.0.20 
pytrends==4.9.2 
pytz==2025.2 
referencing==0.36.2 
requests==2.32.4 
rich==14.0.0 
rpds-py==0.25.1 
sgmllib3k==1.0.0 
shellingham==1.5.4 
six==1.17.0 
smart-open==7.1.0 
smmap==5.0.2 
sniffio==1.3.1 
snscrape==0.7.0.20230622 
soupsieve==2.7 
spacy==3.8.7 
spacy-legacy==3.0.12 
spacy-loggers==1.0.5 
srsly==2.5.1 
starlette==0.46.2 
streamlit==1.46.0 
tenacity==9.1.2 
thinc==8.3.6 
toml==0.10.2 
tornado==6.5.1 
tqdm==4.67.1 
typer==0.16.0 
typing-inspection==0.4.1 
typing_extensions==4.14.0 
tzdata==2025.2 
urllib3==2.5.0 
uvicorn==0.34.3 
wasabi==1.1.3 
watchdog==6.0.0 
weasel==0.4.1 
wrapt==1.17.2 
yarl==1.20.1 